E_USER_WARNING error when png/blur enabled without required functions available
generate error logs when blur is enabled but the required function imagefilter is not available.
likewise if imageFormat="png" but imagepng is not available. 

in development environments, this should catch the developer's attention immediately (display_errors=1 and broken images), whilst on production servers, should at generate some error logs so sysadmins can notice :p

IMO, it's even debatable whether or not image generation should continue, with reduced security, if blur is enabled but imagefilter is not, or if we should just outright cast an exception, because... reduced security! / presumably easier captcha